src/services/auth/requirements.txt
requests==2.28.1
fastapi==0.80.0
uvicorn==0.18.2
pyjwt==2.6.0
sqlalchemy==1.4.44
passlib==1.7.2
bcrypt
email-validator==1.0.5
psycopg2-binary==2.9.5
python-multipart==0.0.5